Update streamlit_app.py
🎯 Bring In Pandas 
Next, we're going to add a new python library called "pandas" into our project .

Remember that when we bring in a new Python library we need to add it to the requirements file, put an import statement into the app file, and reboot the app. Please do those things now. 

When you bring in pandas to the Py file, call it pd in the same way we shorten streamlit to st. This is common in the Python coding world and it makes pandas easier to refer to.